ai-agent-python/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
web3>=6.0,<8
python-dotenv>=1.0.0,<2
